metal_defect_detection/metal_defect_detection/requirements.txt
tensorflow>=2.0
flask
numpy
pillow
